requirements.txt
requests
pyYAML
flask